refactor(web): extract judge ZeroState into pages/judge/ZeroState.tsx (PRD #1049)
